Don't call releaseAll after the Robot finishes running
Initially, this code served to prevent race conditions: if, for example,
the user interrupted a Robot while it was pressing a key, the key would
be left held down rather than released as expected. Calling `releaseAll`
fixed this in a rather heavy-handed way, by releasing *all* keys and
buttons when `runRobot` exits.

But since the affected functions are now implemented in terms of
`bracket`, this approach is redundant -- and we can remove this
unintuitive behavior once and for all.